Threat-Awareness/Anomaly-Detectors/UMU-T4.3-FL-Anomaly-Detection/ai-detection-engine/requirements.txt
Flask==3.1.0
joblib==1.4.2
nats-py==2.9.0
keras==2.15.0
numpy==1.23.5
pandas==1.4.2
requests==2.25.1
scikit_learn==1.5.0
tensorflow==2.15.0
